safe_guard/safe_guard/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
joblib==1.4.2
mysqlclient==2.2.4
numpy==2.0.2
scikit-learn==1.5.2
scipy==1.13.1
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
xgboost==2.1.1
